kimera/container/requirements.txt
kubernetes>=28.1.0
click>=8.1.0
pyyaml>=6.0
rich>=13.7.0
tabulate>=0.9.0
jsonpatch>=1.33
